(update) populated the cards into Intermission Widget
Made it so the cards actually pull from available plots. Currently only server has it so its not replicated and we need to add an index so it takes next available plot. When this is done we need to replicate so the clients have unique cards. After that we will set up a system for selecting the cards and then removing them from the available plots.