(Update) Need to juice up plot ID system
Hard coding wont work. Got it showing up for server but widgets cannot self replicate when a UI is the one filling widget slots. So we need game state to get an array of IDs and somehow get game state to assign plots using this ID. inshallah. For now lewis I unhooked my stuff so you can go to town.