[New] TradeManager Map to Gamestate
I've added a variable in the game state to track which players are associated with a TradeManager when it is spawned. This means the players can reference which Trademanager to respond to. It doesnt have any logic in the game state its just a variable. Please DO NOT DELETE IT HAHA!