[Update] Reworked the trade handshake
Added a session key and swap key which is a Struct containing the two players. Its added to the game state map for trading. Basically, we can now accurately track which 2 players are trading and they can reciprocate. The trade handshake is now working well and firing as expected.

I also updated the cancel trade logic to be a bit cleaner.