(Add & Update) We have a struct now for BP_PLots
Basically redid the system for plots and IDs being added. It works. We can now make all players get cards with real plot IDs on them. Unfortunately, This didnt solve the next problem I have, which is trying to figure out how to actually give players different plot cards. This is the big hertle I need to sort out. Ive been doing this on and off all day and need a break. Inshallah.