(Fix) Fixed Pause input bug
So instead of set player movement stopped, we set input to UI mode only. Then there is a bool that flushes any input so the guy doesnt run forever. or get stuck floating mid jump. Simple disable of access to the player input, and a flush that clears any current inputs.