(fix) Pause issues sorted
Okay we fixed the bug with pause menu that causes players to float, I discovered a new Set Ignore Move Input Option so that when we pause, we simply toggle inputs to stop. Was hard to find this burried in there and AI did not help.

I also discovered that we have an option with SET UI MODE to FLUSH the input so the player doesnt walk around during intermission. And because its INTERMISSION, the player is in UI mode and cant walk around anyways. This is nice since you also cant build or do anything else in UI mode besides select cards, so its nice to have this still as an option.

So now we have TWO options for pausing which is great. Because the new version is also usable for the trade menu if needed or other types of widgets that require players to stop movement. Very scalable.

Also added new input option for TAB which opens a scoreboard widget so player can see who is winning. Currently just says that WEIS wins. Obviously.

The score board menu still allows player movement, since it would feel jank if it stopped you from walking. I would assume players would only look at it for a second or two anyways and would probably want to still do it while running.