(Added) Initializing Permit Deck Completed and bug fix
A Deck of permits is created with a customizable amount of times each card appears in the deck.

Also discovered a bug where if you enter the pause menu, and open the settings, ESCAPE does nothing the first time because its trying to pause an already destroyed menu. BUT the settings are still there so if you press escape again, you have a NEW pause menu and can open a second settings window. The fix was simply adding destruction of the settings menu to our logic. So now when you press ESCAPE, both are destroyed. GO BACK in the settings menu still works because it simply HIDES the pause menu ;)