(update) Fixed card highlighting and set up alternate game starts
We added some sounds to the card selection to help with testing. You can no longer select OR highlight any card past the pre defined limit. It looks clean and the sound is just alright but does the job for basic player feedback.

We also majorly overhauled old and BS functions inside game state and game mode. So game state now automatically will enter the player in the game with 15 permits, and 50k. Each player also gets plots autoassigned with no round system in place. The open sandbox default tester for checking that our systems and buildings work.

We also added a game start variant for ONLY testing the intermission UI and card selection/assignment.

We ALSO added a game start variant for testing BOTH the round timer system, and the intermission.

This change will 200x the time it takes to show someone how the game works, and to test specific features depending on whatever the current need is.

REMOVED - give permits to players logic from game mode is no longer needed and we just do it in game state after making the deck.

REMOVED - insane amount of unused print strings were removed and a lot of logic was cleaned up in game state and intermission related UIs. Lots of comment blocks describing systems so that no work is lost in translation.