(add) Complete update to build system
Build system is exactly the same basically, but now when we build, we enter a build mode, which allows building rotation and saves that rotation. It also stops player movement and goes into a third person camera so you can see more of the building. Exiting the build mode is possible when you place the building successfuly using the saved rotation causing build mode to exit and player movement to return along with the first person camera.

This system is perfect for adding a widget or tooltip for build mode enter and closing it on exit. But for now it just allows us as devs to do the work.

Also had a chance to completely organize all the shit inside of our player controller blueprint so it doesnt look like aids.