(update) Progress bar and bug fix
Added progress bar for press and hold to build, added sounds as well because why not. Temporary. Had to basically fix this weird bug that apparently always existed but was only visable when you had a build mode. But essentially the client now checks with server if the plot has a building on it. If the plot has no building, then we dont enter build mode. The problem was that build mode can be entered since the status of the buildable plot was not replicated. This didnt exist before because we immediately set the plot as built on, so the build function never consumed a permit. it technically built the building for a split second, but denied it and then did not consume permit. Build mode however allows a second or third or fourth building to be made, and consumes permit because it never does this check. But anyways. Fixed that shit.