(update) WIP but working on adding money per placed building.
Read the title. Also i added some simple logic that allows a player to change the currently selected permit by manually updating the building mesh if the build mode is active. and using that new selection as the newly passed selected permit for RPC build and spawn building + remove permit.

Tested this with intermission active and tested with a temporary add balance function and we are 100% able to get the correct building rent of this owned building, even if we switch it halfway through building.